requirements.txt
mlflow==2.5.0
boto3==1.28.2
click==8.1.6
cloudpickle==2.1.0
defusedxml==0.7.1
gitpython==3.1.30
ipython==8.4.0
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
prometheus-client==0.14.1
protobuf==3.20.1
pyopenssl==22.0.0
pytorch-lightning==2.0.6
pyyaml==6.0
rasterio==1.3.4
regex==2022.7.25
requests==2.28.1
torch==1.13.1
torchvision==0.14.1
